+++ src/bin/mash.rs
@@ -16,12 +16,24 @@
 use core::panic::PanicInfo;
 
 use tlenix_core::{
-    consts::EXIT_SUCCESS, io::Console, print, println, process::exit, sleep_loop_forever,
+    consts::EXIT_SUCCESS,
+    eprintln,
+    fs::get_current_working_directory,
+    io::Console,
+    print, println,
+    process::exit,
+    sleep_loop_forever,
+    system::{power_off, reboot},
 };
 
 const MASH_PANIC_TITLE: &str = "mash";
-const PROMPT: &str = ":} ";
+
+const PROMPT_START: &str = "\u{001b}[94mmash\u{001b}[0m";
+const PROMPT_FINISH: &str = "\u{001b}[92;1m:}\u{001b}[0m";
+
 const EXIT_BYTES: &[u8] = b"exit\0";
+const POWEROFF_BYTES: &[u8] = b"poweroff\0";
+const REBOOT_BYTES: &[u8] = b"reboot\0";
 
 const LINE_MAX: usize = 1024;
 
@@ -44,15 +56,25 @@ pub extern "C" fn _start() -> ! {
     let console = Console::open_console().unwrap();
 
     loop {
-        print!("{PROMPT}");
+        prompt();
         let line: [u8; LINE_MAX] = console.read_line().unwrap();
-        // Exit if `exit` is typed
-        if &line[..5] == EXIT_BYTES {
-            exit(EXIT_SUCCESS)
-        }
 
-        // TODO just echo everything back for now
-        if line[0] != 0 {
+        if &line[..5] == EXIT_BYTES {
+            // Exit if `exit` is typed
+            exit(EXIT_SUCCESS)
+        } else if &line[..9] == POWEROFF_BYTES {
+            // Poweroff if `poweroff` is typed
+            let errno = power_off().unwrap_err();
+            // `power_off` should shut down the machine, so if we've made it this far, it's
+            // an error!
+            eprintln!("poweroff fail: {}", errno.as_str());
+        } else if &line[..7] == REBOOT_BYTES {
+            // Reboot if `reboot` is typed
+            let errno = reboot().unwrap_err();
+            // `reboot` should reboot the machine, so if we've made it this far, it's an error!
+            eprintln!("reboot fail: {}", errno.as_str());
+        } else if line[0] != 0 {
+            // TODO just echo everything back for now
             if let Ok(utf8_line) = str::from_utf8(&line) {
                 println!("{utf8_line}");
             } else {
@@ -60,6 +82,25 @@ pub extern "C" fn _start() -> ! {
             }
         }
     }
+}
+
+/// Print the MASH shell prompt.
+fn prompt() {
+    // TODO clean this up
+    let mut cwd_backup: [u8; LINE_MAX] = [0x00_u8; LINE_MAX];
+    cwd_backup[0] = b'?';
+    let cwd_str_backup = "?";
+    let cwd: &[u8; LINE_MAX] = &get_current_working_directory().unwrap_or(cwd_backup);
+    let cwd_str: &str = str::from_utf8(cwd).unwrap_or(cwd_str_backup);
+    let cwd_str_trimmed: &str = cwd_str.trim_end_matches('\0');
+    let basename: &str = cwd_str_trimmed
+        .rsplit_once('/')
+        .map_or(
+            cwd_str_trimmed,
+            |(_, last)| if last.is_empty() { "/" } else { last },
+        );
+
+    print!("{PROMPT_START} {basename} {PROMPT_FINISH} ");
 }
 
 #[panic_handler]