+++ src/data.rs
@@ -11,7 +11,7 @@ const TO_NULL_TERM_STR_ERR_MSG: &str = "input is too long for NullTermStr buffer
 ///
 /// ```
 /// // The byte array itself is 13 bytes long, but with the null terminator it'd be 14.
-/// const MY_NULL_TERM_STR: NullTermStr<14> = nulltermstr!(b"Hello, world!"[14]);
+/// const MY_NULL_TERM_STR: NullTermStr<14> = nulltermstr!(b"Hello, world!");
 /// ```
 #[macro_export]
 macro_rules! nulltermstr {